Move the Unifi Controller to the new tunnel-based stack
This now routes most (more on that later) traffic that the Unifi
Controller has through the remote tunnel in the cloud.  This way, any
remote site that is trying to phone home to the controller hits the VPS
in a datacenter, and never the home internet connection.

Some traffic for the local network hackily goes through the MAN network,
for now.  I've filed #4 to clean that up, but there is some other work
that needs to be handled before that can be tackled.  Search code for
`TODO(#4)` to find all the places that need to be adjusted.